- Implementado la asignacion de un rigen a un host

+++ hosting-service/src/main/java/com/parcial/hosting_service/reposotories/FeatureRepository.java
@@ -4,9 +4,7 @@ import com.parcial.hosting_service.models.Feature;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
 @Repository
-public interface FeatureRepository extends JpaRepository<Feature, Long> {
+public interface FeatureRepository extends JpaRepository<Feature, Integer> {
 
 }
